+++ routes/api.js
@@ -8,11 +8,7 @@ module.exports = (pool) => {
 				p_pub.player_name AS publisher_name,
 				p_ver.player_name AS verifier_name,
 				GROUP_CONCAT(DISTINCT p_all.player_name ORDER BY p_all.player_name SEPARATOR ', ') AS creator_names,
-				GROUP_CONCAT(
-				DISTINCT CONCAT(p_vic.player_name, ':', r.progress) 
-                ORDER BY r.progress DESC, p_vic.player_name ASC 
-                SEPARATOR '|'
-				) AS victor_data
+				GROUP_CONCAT(DISTINCT CONCAT(p_vic.player_name, ':', r.progress) ORDER BY r.progress DESC, p_vic.player_name ASC SEPARATOR '|') AS victor_data
 			FROM demonlist d
 			LEFT JOIN creators c ON d.level_id = c.level_id
 			LEFT JOIN players p_all ON c.player_id = p_all.player_id
@@ -43,9 +39,9 @@ module.exports = (pool) => {
 				const { creator_names, victor_data, ...rest } = row;
 
 				return {
-					level_id: row.level_id,
+					levelID: row.level_id,
 					placement: row.placement,
-					level_name: row.level_name,
+					levelName: row.level_name,
 					creators: creators,
 					publisher: row.publisher_name,
 					verifier: row.verifier_name,
@@ -68,13 +64,13 @@ module.exports = (pool) => {
 		const query = `
 			SELECT
 				s.*,
-				(SELECT GROUP_CONCAT(CONCAT(d2.placement, '::', d2.level_name, '::', p_pub.player_name, '::', r2.progress, '::', r2.is_verifier, '::', d2.points) ORDER BY d2.placement ASC SEPARATOR '|')
+				(SELECT GROUP_CONCAT(CONCAT(d2.level_id, '::', d2.placement, '::', d2.level_name, '::', p_pub.player_name, '::', r2.progress, '::', r2.is_verifier, '::', d2.points, '::', d2.list_percentage_points) ORDER BY d2.placement ASC SEPARATOR '|')
 				FROM records r2
 				JOIN demonlist d2 ON r2.level_id = d2.level_id
 				LEFT JOIN creators c_pub ON d2.level_id = c_pub.level_id AND c_pub.is_publisher = 1
 				LEFT JOIN players p_pub ON c_pub.player_id = p_pub.player_id
 				WHERE r2.player_id = s.player_id AND r2.progress > 0) AS all_records,
-				(SELECT GROUP_CONCAT(CONCAT(d3.placement, '::', d3.level_name, '::', p_pub3.player_name) ORDER BY d3.placement ASC SEPARATOR '|')
+				(SELECT GROUP_CONCAT(CONCAT(d3.level_id, '::', d3.placement, '::', d3.level_name, '::', p_pub3.player_name) ORDER BY d3.placement ASC SEPARATOR '|')
 				FROM creators c
 				JOIN demonlist d3 ON c.level_id = d3.level_id
 				LEFT JOIN creators c_pub3 ON d3.level_id = c_pub3.level_id AND c_pub3.is_publisher = 1
@@ -92,20 +88,22 @@ module.exports = (pool) => {
 
 			const formattedResults = results.map(row => {
 				const allRecords = row.all_records ? row.all_records.split('|').map(item => {
-					const [placement, name, publisher, progress, isVerifier, points] = item.split('::');
+					const [levelID, placement, name, publisher, progress, isVerifier, points, listPercentagePoints] = item.split('::');
 					return {
+						levelID: parseInt(levelID),
 						placement: parseInt(placement),
 						name: name,
 						publisher: publisher,
 						progress: parseInt(progress),
 						isVerifier: isVerifier === '1',
-						points: parseInt(points)
+						points: parseInt(points),
+						listPercentagePoints: parseInt(listPercentagePoints),
 					};
 				}) : [];
 
 				const created = row.levels_created ? row.levels_created.split('|').map(item => {
-					const [placement, name, publisher] = item.split('::');
-						return { placement: parseInt(placement), name: name, publisher: publisher };
+					const [levelID, placement, name, publisher] = item.split('::');
+						return { levelID: parseInt(levelID), placement: parseInt(placement), name: name, publisher: publisher };
 				}) : [];
 
 				return {
@@ -115,13 +113,13 @@ module.exports = (pool) => {
 					points: parseInt(row.points) || 0,
 					levelsVerified: allRecords
 						.filter(r => r.isVerifier)
-						.map(({ placement, name, publisher, points }) => ({ placement, name, publisher, points })),
+						.map(({ levelID, placement, name, publisher, points }) => ({ levelID, placement, name, publisher, points })),
 					levelsCompleted: allRecords
 						.filter(r => !r.isVerifier && r.progress === 100)
-						.map(({ placement, name, publisher, points }) => ({ placement, name, publisher, points })),
+						.map(({ levelID, placement, name, publisher, points }) => ({ levelID, placement, name, publisher, points })),
 					progressOn: allRecords
 						.filter(r => !r.isVerifier && r.progress < 100)
-						.map(({ placement, name, publisher, progress, points }) => ({ placement, name, publisher, progress, points })),
+						.map(({ levelID, placement, name, publisher, progress, points, listPercentagePoints}) => ({ levelID, placement, name, publisher, progress, points, listPercentagePoints })),
 					levelsCreated: created
 				};
 			});